Rename Docker tests from ServletIT to DockerIT and JettyServerFactory-based Spec to ServletIT spec
